+++ src/hooks/index.ts
@@ -1,8 +1,10 @@
 import {useDialog} from './useDialog';
 import {usePrevious} from './usePrevious';
 import {useDebounce, useDebouncedCallback} from 'use-debounce';
+import {useDropdownOptions} from './useDropdownOptions';
 import {useLocalStorage} from './useLocalStorage';
 import {useIsomorphicLocalStorage} from './useIsomorphicLocalStorage';
+import {useClipboard} from './useClipboard';
 import {useToggle} from './useToggle';
 
 export {
@@ -10,7 +12,9 @@ export {
   usePrevious,
   useDebounce,
   useDebouncedCallback,
+  useDropdownOptions,
   useLocalStorage,
   useIsomorphicLocalStorage,
+  useClipboard,
   useToggle,
 };